+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - SignedIndexArray.equals()
 - class IndexArray<EL>, same as SignedIndexArray but for non-negative indices.
 - Added KVComponent interface for Map1, Map2, Map3, IndexArray and SignedIndexArray.
+- Added MultiContainer.
 
 ## [1.14.0] - 2025-10-19
 ### Added